fix: lista 2 errors

+++ Banco/Banco/Banco/ContaEspecial.cpp
@@ -1,7 +1,6 @@
 #include "ContaEspecial.h"
 
-void ContaEspecial::debitar(int valor) {
-
+void ContaEspecial::debitar(double valor) {
     if (valor <= (getSaldo()+limiteCredito))
         setSaldo(getSaldo() - valor);
 }
